+++ app/src/main/java/org/simple/clinic/allpatientsinfacility/AllPatientsInFacilityView.kt
@@ -15,7 +15,7 @@ import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityListItem.AllPatientsInFacilityListItemCallback
 import org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityListItem.Event.SearchResultClicked
 import org.simple.clinic.databinding.ListAllpatientsinfacilityFacilityHeaderBinding
-import org.simple.clinic.databinding.ListPatientSearchBinding
+import org.simple.clinic.databinding.ListPatientSearchOldBinding
 import org.simple.clinic.databinding.ViewAllpatientsinfacilityBinding
 import org.simple.clinic.di.injector
 import org.simple.clinic.facility.FacilityRepository
@@ -40,8 +40,8 @@ class AllPatientsInFacilityView(
             R.layout.list_allpatientsinfacility_facility_header to { layoutInflater, parent ->
               ListAllpatientsinfacilityFacilityHeaderBinding.inflate(layoutInflater, parent, false)
             },
-            R.layout.list_patient_search to { layoutInflater, parent ->
-              ListPatientSearchBinding.inflate(layoutInflater, parent, false)
+            R.layout.list_patient_search_old to { layoutInflater, parent ->
+              ListPatientSearchOldBinding.inflate(layoutInflater, parent, false)
             }
         ),
         locale = locale